requirements.txt
# JEE-Helper Physics Tutor - Requirements

# Google AI & ADK (install first - they define dependency versions)
google-adk==0.1.0
google-genai>=1.9.0

# Core Framework (compatible with google-adk)
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# CORS support
python-multipart>=0.0.12

# Optional: For development/testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
